requirements.txt
absl-py==1.3.0
ai-safety-gridworlds @ https://github.com/timorl/ai-safety-gridworlds/tarball/master
antlr4-python3-runtime==4.9.3
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.6
cycler==0.11.0
docker-pycreds==0.4.0
flatbuffers==2.0.7
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
gym==0.24.1
gym-notices==0.0.8
hydra-core==1.2.0
idna==3.4
importlib-metadata==5.1.0
jax==0.3.13
jaxlib==0.3.10
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==22.0
pathtools==0.1.2
Pillow==9.3.0
promise==2.3
protobuf==4.21.11
psutil==5.9.4
pycolab==1.2
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
safe-grid-gym @ git+https://github.com/david-lindner/safe-grid-gym.git@f521e906735bb712a0217f491f269da2890c0f00
scipy==1.9.3
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
torch-tb-profiler==0.4.0
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.13.6
zipp==3.11.0
